% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ecodata_indicator_stats.R
\name{collect_ecodata_indicator_stats}
\alias{collect_ecodata_indicator_stats}
\title{Collect indicator stats and create dataset}
\usage{
collect_ecodata_indicator_stats(datalist = NULL, outfile = NULL)
}
\arguments{
\item{datalist}{Optional, vector of ecodata indicator dataset names to be summarized in
ecodata::dataset format. Default is all datasets in the package
#'@param outfile Optional; name of output datafile without extension (.rds format), defaults to "ecodatadat"}
}
\value{
a dataframe with basic stats calculated from the indicator(s). Columns:
\itemize{
\item{\code{Indicator}, name of the variable summarized from the ecodata indicator dataset}
\item{\code{Dataset}, the ecodata indicator dataset name = "ecodata name" collected from catalog}
\item{\code{Units}, the units of the indicator}
\item{\code{EPU}, the Ecosystem Production Unit where the indicator is measured}
\item{\code{Varname}, basic indicator statistics; StartYear is the first year of the time series,
EndYear is the last year of the time series, EstYrs is the total possible years between the start
and end years, ActualYrs is the count of years to determine if any are missing, TSMean is the time
series mean value, TSmin is the time series minimum value, TSmax is the time series maximum value,
YrTSmin is the year of the time series minimum value, YrTSmax is the year of the time series maximum value}
\item{\code{Value}, value of the statistic for the indicator}
}
}
\description{
Apply \code{ecodata_indicator_stats} to all or a subset of ecodata package datasets and save collated dataset
}
\examples{
 datalist <- c("ecodata::trans_dates",
 "ecodata::SAV",
 "ecodata::zooplankton_index")
 outfile <- "testecodat"
 collect_ecodata_indicator_stats(datalist, outfile)

}
