% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ecodata_techdoc_data.R
\name{collect_ecodata_techdoc_data}
\alias{collect_ecodata_techdoc_data}
\title{Collect ecodata techdoc data and create dataset}
\usage{
collect_ecodata_techdoc_data(urllist = NULL, outfile = NULL)
}
\arguments{
\item{urllist}{Optional; the urls of the tech-doc page to be scraped, defaults to all pages}

\item{outfile}{Optional; name of output datafile without extension (.rds format), defaults to "techdocdat"}
}
\value{
saves an R dataset (.rds) with variables from the FieldList and values from the tech-doc webpage. Columns:
\itemize{
\item{\code{Indicator}, name of the indicator extracted from tech-doc page title}
\item{\code{Source}, URL of the tech-doc page}
\item{\code{Varname}, variable names corresponding to matches from the FieldList}
\item{\code{Value}, text for each variable name from the tech-doc page}
}
}
\description{
Apply the \code{scrape_ecodata_techdoc} function to all or a subset of tech-doc pages and save the dataset
}
\examples{
 urllist <- c("https://noaa-edab.github.io/tech-doc/trans_dates.html",
 "https://noaa-edab.github.io/tech-doc/SAV.html",
 "https://noaa-edab.github.io/tech-doc/zooplankton_index.html")
 outfile <- "testdat"
 collect_ecodata_techdoc_data(urllist, outfile)

}
