% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_marea_indicator_stats.R
\name{collect_marea_indicator_stats}
\alias{collect_marea_indicator_stats}
\title{Collect indicator stats and create dataset}
\usage{
collect_marea_indicator_stats(datalist = NULL, outfile = NULL)
}
\arguments{
\item{datalist}{Optional, vector of marea indicator dataset names to be summarized in
ecodata::dataset format. Default is all datasets in the package
#'@param outfile Optional; name of output datafile without extension (.rds format), defaults to "mareadat"}
}
\value{
a dataframe with basic stats calculated from the indicator(s). Columns:
\itemize{
\item{\code{Indicator}, Indicator name}
\item{\code{Dataset}, the marea indicator dataset name}
\item{\code{Var}, name of the indicator variable summarized from the marea indicator dataset}
\item{\code{Source}, the source of the indicator}
\item{\code{Units}, the units of the indicator}
\item{\code{Time}, periodicity of the indicator: annual or annual and monthly}
\item{\code{Space}, spatial description}
\item{\code{EPU}, the Ecosystem Production Unit or Region where the indicator is measured}
\item{\code{Varname}, basic indicator statistics; StartYear is the first year of the time series,
EndYear is the last year of the time series, EstYrs is the total possible years between the start
and end years, ActualYrs is the count of years to determine if any are missing, TSMean is the time
series mean value, TSmin is the time series minimum value, TSmax is the time series maximum value,
YrTSmin is the year of the time series minimum value, YrTSmax is the year of the time series maximum value}
\item{\code{Value}, value of the statistic for the indicator}
}
}
\description{
Apply \code{marea_indicator_stats} to all or a subset of marea package datasets and save collated dataset
}
\examples{
 datalist <- c("amo",
 "ao",
 "nao")
 outfile <- "testmarea"
 collect_marea_indicator_stats(datalist, outfile)

}
