% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_smart_ratings.R
\name{combine_smart_ratings}
\alias{combine_smart_ratings}
\title{Combine individual indicator SMART ratings from rmd rendering and create dataset}
\usage{
combine_smart_ratings(csvlist = NULL, outfile = NULL)
}
\arguments{
\item{csvlist}{Optional; list of smart rating csvs created by rendering indicator rmd files, defaults to all csvs in \code{data-raw/indratings}}

\item{outfile}{Optional; name of output datafile without extension (.rds format), defaults to "smartratingdat"}
}
\value{
saves an R dataset (.rds) with smart elements, attributes, and rankings. Columns:
\itemize{
\item{\code{Category}, Character string; indicator category from the SMART page}
\item{\code{Indicator}, Character string; name of the indicator}
\item{\code{Element}, Character string; SMART element: Specific, Measurable, Achievable, Relevant, Time-bound, or All}
\item{\code{Attribute}, Character string; attributes of each SMART element, or denoting element summary ("ElementRating")
or summary across all elements ("OverallRating")}
\item{\code{Rating}, Numeric; rating evaluating how well the indicator meets attribute criteria; range from 0 does not meet to 1 fully meets.
Ratings for each element "ElementRating" are an average across all attributes within an element. "OverallRating" is an average across
all "ElementRating".}
}
}
\description{
SMART ratings include those for individual attributes as well as summary ratings by attribute and across all SMART elements
}
\examples{
 csvlist <- c(here::here("data-raw/indratings/Gulf Stream Index_ratingsumm.csv"),
 here::here("data-raw/indratings/Forage Fish Index_ratingsumm.csv"),
 here::here("data-raw/indratings/Harbor Porpoise Bycatch_ratingsumm.csv"))
 outfile <- "test_smartratingdat"
 combine_smart_ratings(csvlist, outfile)

}
