% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template_SOE.R
\name{create_template_SOE}
\alias{create_template_SOE}
\title{Create a SMART indicator bookdown template for an indicator of interest from the NEFSC SOE.}
\usage{
create_template_SOE(
  indicator_name,
  indicator_varname = NULL,
  template = here::here("templates/SMART_template_SOE.rmd"),
  output_dir = here::here("docs"),
  send_to_google_doc = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{indicator_name}{Character string. An indicator name (e.g. "Phytoplankton")}

\item{indicator_varname}{Character string. The variable name for indicators with multiple variables (e.g. "ANNUAL_PPD_MEDIAN"), defaults to all indicator variables, which can be a lot}

\item{template}{Character string. The path to the desired template, defaults to "templates/SMART_template.rmd"}

\item{output_dir}{Character string. The output directory for the compiled bookdown HTML document and supporting files. Default "docs" folder in current working directory}

\item{send_to_google_doc}{Logical. Defaults to FALSE. If TRUE, the generated template will render a google document into the chosen output directory for hand editing.
If FALSE, the template .Rmd file will be generated in the directory output_dir/stock-name folder}

\item{overwrite}{Logical. Defaults to FALSE. If TRUE, output will overwrite any existing template or google doc for chosen species.}
}
\value{
A .Rmd file and/or google document populated with figures
}
\description{
Inspired by NOAA-EDAB Ecosystem Context for Stock Assessment workflow
original code: https://github.com/NOAA-EDAB/ECSA/blob/master/R/create_template.R
}
\examples{

create_template_SOE(indicator_name = "Phytoplankton", indicator_varname = "ANNUAL_PPD_MEDIAN",
overwrite = F, output_dir = getwd(), send_to_google_doc = T)

}
