% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata_indicator_stats.R
\name{ecodata_indicator_stats}
\alias{ecodata_indicator_stats}
\title{Summarize and visualize ecodata indicators}
\usage{
ecodata_indicator_stats(dataset)
}
\arguments{
\item{dataset}{The ecodata indicator dataset name in ecodata::dataset format (may contain one or multiple indicator variables)}
}
\value{
a dataframe with variables from the StatList and values calculated from the indicator(s). Columns:
\itemize{
\item{\code{Indicator}, name of the variable summarized from the ecodata indicator dataset}
\item{\code{Dataset}, the ecodata indicator dataset name = "ecodata name" collected from catalog}
\item{\code{Units}, the units of the indicator}
\item{\code{EPU}, the Ecosystem Production Unit where the indicator is measured; }
\item{\code{Varname}, basic indicator statistics; StartYear is the first year of the time series,
EndYear is the last year of the time series, EstYrs is the total possible years between the start
and end years, ActualYrs is the count of years to determine if any are missing, TSMean is the time
series mean value, TSmin is the time series minimum value, TSmax is the time series maximum value,
YrTSmin is the year of the time series minimum value, YrTSmax is the year of the time series maximum value}
\item{\code{Value}, value of the statistic for the indicator}
}
}
\description{
Retrieve attributes of an ecodata indicator and return as a dataframe, and plot the indicator.
}
\examples{
 ecodata_indicator_stats(dataset = ecodata::forage_index)

}
