% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_indiator_templates.R
\name{fill_indicator_templates}
\alias{fill_indicator_templates}
\title{Create and fill .Rmd templates for multiple indicators}
\usage{
fill_indicator_templates(indlist = NULL, overwrite = FALSE)
}
\arguments{
\item{indlist}{Optional; the names of indicators to create templates for, defaults to all in the catalog dataset created by \code{collect_ecodata_catalog_data}}

\item{overwrite}{Logical. Defaults to FALSE. If TRUE, output will overwrite any existing template or google doc for chosen species.}
}
\value{
.Rmd templates for each indicator in the indlist with the naming convention \code{indicator_name.Rmd}.
}
\description{
Apply the \code{create_template} function to all or a subset of ecosystem indicators
}
\examples{
 indlist <- c("Transition Dates","Submerged Aquatic Vegetation","Zooplankton Indices")
 fill_indicator_templates(indlist, overwrite = TRUE)

}
