% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_marea_templates.R
\name{fill_marea_templates}
\alias{fill_marea_templates}
\title{Create and fill .Rmd templates for multiple indicators}
\usage{
fill_marea_templates(indlist = NULL, overwrite = FALSE)
}
\arguments{
\item{indlist}{Optional; the names of indicators to create templates for, defaults to all in the marea dataset created by \code{collect_marea_indicator_stats}}

\item{overwrite}{Logical. Defaults to FALSE. If TRUE, output will overwrite any existing template or google doc for chosen species.}
}
\value{
.Rmd templates for each indicator in the indlist with the naming convention \code{indicator_name.Rmd}.
}
\description{
Apply the \code{create_template_marea} function to all or a subset of marea ecosystem indicators
}
\examples{
 indlist <- c("Atlantic MultiDecadal Oscillation Index", "Arctic Oscillation Index", "North Atlantic Oscillation Index")
 fill_marea_templates(indlist, overwrite = TRUE)

}
