% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_to_filledrmd.R
\name{gdoc_to_filledrmd}
\alias{gdoc_to_filledrmd}
\title{#' Create filled SMART template Rmd from an edited google doc for further processing.}
\usage{
gdoc_to_filledrmd(
  indicator_name,
  input_dir = here::here("drafts"),
  output_dir = here::here("docs"),
  overwrite = FALSE
)
}
\arguments{
\item{indicator_name}{Character string. An indicator name (e.g. "Phytoplankton")}

\item{input_dir}{Character string. The original draft rmd that produced the unedited gdoc is in this directory. Defaults to "drafts" folder in current working directory}

\item{output_dir}{Character string. The output directory for the compiled bookdown HTML document and supporting files. Default "docs" folder in current working directory}

\item{overwrite}{Logical. If TRUE, output will overwrite any existing template for chosen species.}
}
\value{
An .Rmd file populated with text and figures
}
\description{
#' Create filled SMART template Rmd from an edited google doc for further processing.
}
\details{
Inspired by ECSA merge_to_bookdown https://github.com/NOAA-EDAB/ECSA/blob/master/R/merge_to_bookdown.R
But only takes a filled gdoc back to rmd to use render_smartind and renderall for the full book
}
\examples{

create_template(indicator_name = "smooth-dogfish", overwrite = T, output_dir = getwd())

}
