% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marea_indicator_stats.R
\name{marea_indicator_stats}
\alias{marea_indicator_stats}
\title{Summarize and visualize marea indicators}
\usage{
marea_indicator_stats(dataset)
}
\arguments{
\item{dataset}{The marea indicator dataset name in ecodata::dataset format (may contain one or multiple indicator variables)}
}
\value{
a dataframe with variables from the StatList and values calculated from the indicator(s). Columns:
\itemize{
\item{\code{Indicator}, Indicator name}
\item{\code{Dataset}, the marea indicator dataset name}
\item{\code{Var}, name of the indicator variable summarized from the marea indicator dataset}
\item{\code{Source}, the source of the indicator}
\item{\code{Units}, the units of the indicator}
\item{\code{Time}, periodicity of the indicator: annual or annual and monthly}
\item{\code{Space}, spatial description}
\item{\code{EPU}, the Ecosystem Production Unit or Region where the indicator is measured}
\item{\code{Varname}, basic indicator statistics; StartYear is the first year of the time series,
EndYear is the last year of the time series, EstYrs is the total possible years between the start
and end years, ActualYrs is the count of years to determine if any are missing, TSMean is the time
series mean value, TSmin is the time series minimum value, TSmax is the time series maximum value,
YrTSmin is the year of the time series minimum value, YrTSmax is the year of the time series maximum value}
\item{\code{Value}, value of the statistic for the indicator}
}
}
\description{
Retrieve attributes of a marea indicator and return as a dataframe, and plot the indicator.
}
\examples{
 marea_indicator_stats(dataset = "amo")

}
