% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_smartind.R
\name{render_smartind}
\alias{render_smartind}
\title{Render single .Rmd template using bookdown format}
\usage{
render_smartind(input_file)
}
\arguments{
\item{input_file}{file name to be rendered,  in \code{here::here("docs")}}
}
\value{
.html rendered using bookdown, or whatever format is first in _output.yml
}
\description{
Render single .Rmd template using bookdown format
}
\examples{
 render_smartind("Aquaculture-Production.rmd")

}
