% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderall.R
\name{renderall}
\alias{renderall}
\title{Render all .Rmd templates in a given directory}
\usage{
renderall(rmddir = NULL, output = "html", overwrite = FALSE)
}
\arguments{
\item{output}{Optional; Defaults to "html". Other options are "all", "pdf" and "word" but your mileage will vary.}

\item{overwrite}{Logical. Defaults to FALSE. If TRUE, output will overwrite any existing html.}

\item{moddir}{Optional; the path to the dirctory containing .Rmd files to be rendered, defaults to \code{here::here("docs")}}
}
\value{
.Rmd templates for each indicator in the indlist with the naming convention \code{indicator_name.Rmd}.
}
\description{
Render all .Rmd templates in a given directory
}
\examples{
 rmddir <- here::here("test")
 renderall(rmddir, overwrite = TRUE)

}
