% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ecodata_techdoc.R
\name{scrape_ecodata_techdoc}
\alias{scrape_ecodata_techdoc}
\title{Scrape ecodata tech-doc page}
\usage{
scrape_ecodata_techdoc(url, FieldList = NULL)
}
\arguments{
\item{url}{The url of the tech-doc page to be scraped}

\item{FieldList}{Optional, A character vector of fields to be scraped; must match spelling.
The default FieldList is c("Description", "Data steward", "Public availability statement",
"Data sources", "Data extraction", "Data analysis","catalog link")}
}
\value{
a dataframe with variables from the FieldList and values from the tech-doc webpage. Columns:
\itemize{
\item{\code{Indicator}, name of the indicator extracted from tech-doc page title}
\item{\code{Source}, URL of the tech-doc page}
\item{\code{Varname}, variable names corresponding to matches from the FieldList}
\item{\code{Value}, text for each variable name from the tech-doc page}
}
}
\description{
Pulls selected fields, paragraphs, and images from the NOAA-EDAB Tech-doc website
and returns a dataframe of the .
Code had to be tailored to this website because not all fields come under standard headers.
}
