% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_smartinddata.R
\name{scrape_smartinddata}
\alias{scrape_smartinddata}
\title{Scrape smart indicator data pages to make a dataset}
\usage{
scrape_smartinddata(url)
}
\arguments{
\item{url}{The url of the SMART indicator page to be scraped}

\item{FieldList}{Optional, A character vector of fields to be scraped; must match spelling.
The default FieldList is c()}
}
\value{
a dataframe with variables from the FieldList and values from the SMART indicator webpage. Columns:
\itemize{
\item{\code{Indicator}, name of the indicator extracted from SMART indicator page title}
\item{\code{Source}, URL of the SMART indicator page}
\item{\code{Varname}, variable names corresponding to matches from the FieldList}
\item{\code{Value}, text for each variable name from the SMART indicator page}
}
}
\description{
Pulls selected fields, paragraphs, and images from the SMART indicator reports website
and returns a dataframe of the information for each indicator.
This dataset is the input for the searchable shiny app/dashboard
}
\examples{
 url <- "https://mafmc.github.io/SMARTindicatorsData/Cold-Pool-Index.html"
 scrape_smartinddata(url)

}
