% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ecodata_catalog_data.R
\name{collect_ecodata_catalog_data}
\alias{collect_ecodata_catalog_data}
\title{Collect catalog data and create dataset}
\usage{
collect_ecodata_catalog_data(urllist = NULL, outfile = NULL)
}
\arguments{
\item{urllist}{Optional; the urls of the catalog page to be scraped, defaults to all pages}

\item{outfile}{Optional; name of output datafile without extension (.rds format), defaults to "catalogdat"}
}
\value{
a dataframe with variables from the FieldList and values from the catalog webpage. Columns:
\itemize{
\item{\code{Indicator}, name of the indicator extracted from catalog page title}
\item{\code{Source}, URL of the catalog page}
\item{\code{Varname}, variable names corresponding to matches from the FieldList}
\item{\code{Value}, text for each variable name from the catalog page}
}
}
\description{
Apply the scrape_ecodata_catalog function to all or a subset of catalog pages and create a dataset
}
