% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderall.R
\name{renderall}
\alias{renderall}
\title{Render all .Rmd templates in a given directory using bookdown}
\usage{
renderall(rmddir = NULL, overwrite = FALSE)
}
\arguments{
\item{overwrite}{Logical. Defaults to FALSE. If TRUE, output will overwrite any existing html.}

\item{moddir}{Optional; the path to the dirctory containing .Rmd files to be rendered, defaults to \code{here::here("docs")}}
}
\value{
.html for each indicator in the indlist with the naming convention \code{indicator_name.html}.
}
\description{
Render all .Rmd templates in a given directory using bookdown
}
\examples{
 rmddir <- here::here("test")
 renderall(rmddir, overwrite = TRUE)

}
